+++ tests/integration/cases/dict_mixed_scalars/Go_combined.go
@@ -1,0 +1,13 @@
+package main
+
+func main() {
+my_data := map[string]any{
+	"a": 1,
+	"b": "x",
+}
+my_data = map[string]any{
+	"a": 1,
+	"b": "x",
+}
+_ = my_data
+}