+++ tests/integration/cases/dict_mixed_scalars/V.v
@@ -1,0 +1,8 @@
+
+fn main() {
+	my_data := {
+		'a': 1,
+		'b': 'x',
+	}
+	_ = my_data
+}