+++ tests/integration/test_integration.py
@@ -1548,6 +1548,7 @@ def _build_variant_cases() -> list[_VariantCase]:
         (_build_constructor_name_variants(), "simple_dict", ""),
         (type_hints_cross, "bool_list", ""),
         (type_hints_cross, "float_list", ""),
+        (heterogeneous_strategy, "dict_mixed_scalars", ""),
         (heterogeneous_strategy, "mixed_type_dicts_in_sequence", ""),
         (heterogeneous_strategy, "nested_mixed_types", "_sibling"),
         (heterogeneous_strategy, "nested_mixed_inner", "_inner"),